+++ test/components/DayPickerSingleDateController_spec.jsx
@@ -14,6 +14,7 @@ import * as isDayVisible from '../../src/utils/isDayVisible';
 import getVisibleDays from '../../src/utils/getVisibleDays';
 
 import { VERTICAL_SCROLLABLE } from '../../src/constants';
+import DayPickerRangeController from '../../src/components/DayPickerRangeController';
 
 // Set to noon to mimic how days in the picker are configured internally
 const today = moment().startOf('day').hours(12);
@@ -969,6 +970,74 @@ describe('DayPickerSingleDateController', () => {
       expect(onNextMonthClickStub.callCount).to.equal(1);
       expect(onNextMonthClickStub.firstCall.args[0].year()).to.equal(newMonth.year());
       expect(onNextMonthClickStub.firstCall.args[0].month()).to.equal(newMonth.month());
+    });
+  });
+
+  describe('#onMonchChange', () => {
+    it('sets disableNext as false when maxDate is in visible month and a past month is selected', () => {
+      const wrapper = shallow((
+        <DayPickerRangeController
+          onDatesChange={sinon.stub()}
+          onFocusChange={sinon.stub()}
+          maxDate={today}
+        />
+      ));
+      wrapper.setState({
+        currentMonth: today,
+      });
+      expect(wrapper.state()).to.have.property('disableNext', true);
+      wrapper.instance().onMonthChange(today.clone().subtract(1, 'month'));
+      expect(wrapper.state()).to.have.property('disableNext', false);
+    });
+
+    it('sets disablePrev as false when minDate is in visible month and a future month is selected', () => {
+      const wrapper = shallow((
+        <DayPickerRangeController
+          onDatesChange={sinon.stub()}
+          onFocusChange={sinon.stub()}
+          minDate={today}
+        />
+      ));
+      wrapper.setState({
+        currentMonth: today,
+      });
+      expect(wrapper.state()).to.have.property('disablePrev', true);
+      wrapper.instance().onMonthChange(today.clone().add(1, 'month'));
+      expect(wrapper.state()).to.have.property('disablePrev', false);
+    });
+  });
+
+  describe('#onYearChange', () => {
+    it('sets disableNext as false when maxDate is in visible month and a past year is selected', () => {
+      const wrapper = shallow((
+        <DayPickerRangeController
+          onDatesChange={sinon.stub()}
+          onFocusChange={sinon.stub()}
+          maxDate={today}
+        />
+      ));
+      wrapper.setState({
+        currentMonth: today,
+      });
+      expect(wrapper.state()).to.have.property('disableNext', true);
+      wrapper.instance().onYearChange(today.clone().subtract(1, 'year'));
+      expect(wrapper.state()).to.have.property('disableNext', false);
+    });
+
+    it('sets disablePrev as false when minDate is in visible month and a future year is selected', () => {
+      const wrapper = shallow((
+        <DayPickerRangeController
+          onDatesChange={sinon.stub()}
+          onFocusChange={sinon.stub()}
+          minDate={today}
+        />
+      ));
+      wrapper.setState({
+        currentMonth: today,
+      });
+      expect(wrapper.state()).to.have.property('disablePrev', true);
+      wrapper.instance().onYearChange(today.clone().add(1, 'year'));
+      expect(wrapper.state()).to.have.property('disablePrev', false);
     });
   });